+++ packages/components-angular/src/app/pages/text.component.ts
@@ -99,11 +99,11 @@ import { Component } from '@angular/core';
       </p-text>
     </div>
 
-    <div class="playground" title="should show text with a link and bold text as children">
+    <div class="playground" title="should show text with a link, bold text and italic text as children">
       <p-text>
         <span
           >Lorem ipsum dolor sit amet <a href="#">linked text</a> et, <b>bold text</b> &amp;
-          <strong>strong text</strong></span
+          <strong>strong text</strong>, <em>emphasized text</em> & <i>italic text</i></span
         >
       </p-text>
     </div>