+++ packages/components/src/components/feedback/spinner/spinner.tsx
@@ -1,6 +1,8 @@
-import { JSX, Component, Prop, h } from '@stencil/core';
+import { JSX, Component, Prop, h, Watch } from '@stencil/core';
+import type { Theme } from '../../../types';
+import type { SpinnerSize } from './spinner-utils';
+import { verifySpinnerSize } from './spinner-utils';
 import { isDark, mapBreakpointPropToPrefixedClasses, prefix } from '../../../utils';
-import type { BreakpointCustomizable, Theme } from '../../../types';
 
 @Component({
   tag: 'p-spinner',
@@ -9,10 +11,19 @@ import type { BreakpointCustomizable, Theme } from '../../../types';
 })
 export class Spinner {
   /** Size of the spinner. */
-  @Prop() public size?: BreakpointCustomizable<'small' | 'medium' | 'large' | 'inherit'> = 'small';
+  @Prop() public size?: SpinnerSize = 'small';
 
   /** Adapts the spinner color depending on the theme. */
   @Prop() public theme?: Theme = 'light';
+
+  @Watch('size')
+  public watchSizeHandler(newValue: SpinnerSize): void {
+    verifySpinnerSize(newValue);
+  }
+
+  public componentWillLoad(): void {
+    verifySpinnerSize(this.size);
+  }
 
   public render(): JSX.Element {
     const spinnerClasses = {