+++ packages/components/src/components/form/select-wrapper/select-wrapper.tsx
@@ -322,7 +322,7 @@ export class SelectWrapper {
   }
 
   private handleClickOutside = (e: MouseEvent): void => {
-    if (!this.host.contains(e.target as HTMLElement)) {
+    if (!e.composedPath().includes(this.host)) {
       this.handleVisibilityOfFakeOptionList('hide');
     }
   };
@@ -505,9 +505,7 @@ export class SelectWrapper {
       this.filterInput.value = '';
       this.searchString = '';
       this.filterHasResults = true;
-      if (document.activeElement !== this.filterInput) {
-        this.filterInput.focus();
-      }
+      this.filterInput.focus();
     } else {
       if (document.activeElement !== this.select) {
         this.select.focus();