+++ packages/components-react/src/pages/Spinner.tsx
@@ -21,6 +21,10 @@ export const SpinnerPage = (): JSX.Element => {
       <div className="playground light" title="should show spinner in different sizes on different viewports">
         <PSpinner size={{ base: 'small', m: 'medium', l: 'large' }} />
       </div>
+
+      <div className="playground light" title="should show spinner small when size is undefined">
+        <PSpinner size={undefined} />
+      </div>
     </>
   );
 };