+++ packages/components/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Fixed
+- Usage of `Select Wrapper` within custom elements
+
 ### [2.0.0-rc.4] - 2021-03-01
 
 #### Changed