+++ packages/components/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Changed
+- Added possibility to set background color of `Content Wrapper` to `white`
+
 #### Fixed
 - Usage of `Select Wrapper` within custom elements