+++ packages/components/CHANGELOG.md
@@ -8,8 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Added
+- Configurable background color of `Content Wrapper`
+- `italic` font-style in `Text` is now overridden with `normal`
+
 #### Changed
 - Make `Grid` and `Grid Item` use ShadowDom
+
+#### Fixed
+- Usage of `Select Wrapper` within custom elements
+- A bug that caused `Spinner` to be displayed in a wrong size
 
 ### [2.0.0-rc.4] - 2021-03-01